pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "agentic-problem-solver"
version = "0.1.0"
authors = [
    { name="Anton Tegnelöv", email="anton.tegnelov@gmail.com" },
]
description = "An AI-powered problem-solving system"
readme = "README.md"
requires-python = ">=3.12"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "click",
    "python-dotenv",
    "langchain",
    "langchain-core",
    "google-generativeai",
    "pytest",
    "ruff",
]

[project.scripts]
APS = "src.cli.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
exclude = ["tests*"]

[tool.ruff]
target-version = "py312"
line-length = 88

[tool.ruff.lint]
select = ["ALL"]
extend-safe-fixes = ["ALL"]
ignore = ["D203", "D213"]